removes unnecessary files
