--- auth/cloud-client-temp/authenticate_implicit_with_adc.py
@@ -1,46 +1,0 @@
-# Copyright 2022 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# [START auth_cloud_implicit_adc]
-
-from google.cloud import storage
-
-
-def authenticate_implicit_with_adc(project_id: str = "your-google-cloud-project-id") -> None:
-    """
-    When interacting with Google Cloud Client libraries, the library can auto-detect the
-    credentials to use.
-
-    // TODO(Developer):
-    //  1. Before running this sample,
-    //  set up ADC as described in https://cloud.google.com/docs/authentication/external/set-up-adc
-    //  2. Replace the project variable.
-    //  3. Make sure that the user account or service account that you are using
-    //  has the required permissions. For this sample, you must have "storage.buckets.list".
-    Args:
-        project_id: The project id of your Google Cloud project.
-    """
-
-    # This snippet demonstrates how to list buckets.
-    # *NOTE*: Replace the client created below with the client required for your application.
-    # Note that the credentials are not specified when constructing the client.
-    # Hence, the client library will look for credentials using ADC.
-    storage_client = storage.Client(project=project_id)
-    buckets = storage_client.list_buckets()
-    print("Buckets:")
-    for bucket in buckets:
-        print(bucket.name)
-    print("Listed all storage buckets.")
-
-# [END auth_cloud_implicit_adc]
