--- auth/cloud-client/requirements-test.txt
@@ -1,1 +1,0 @@
-pytest==8.2.0
