+++ auth/cloud-client/requirements.txt
@@ -1,2 +1,8 @@
-google-cloud-storage==2.9.0; python_version < '3.7'
-google-cloud-storage==2.9.0; python_version > '3.6'
+google-cloud-compute==1.42.0
+google-cloud-storage==3.8.0
+google-auth==2.47.0
+pytest===8.4.2; python_version == '3.9'
+pytest==9.0.2; python_version > '3.9'
+boto3>=1.26.0
+requests==2.32.5
+python-dotenv==1.2.1
